Update and rename tplink-smartplug.py to tplink_smartplug.py
